TT-1304: Extract common code for redirect to idp controllers
Author: @NasAshrafThoughtworks
