+++ app/controllers/redirect_to_idp_question_loa2_controller.rb
@@ -1,34 +1,21 @@
 class RedirectToIdpQuestionLoa2Controller < ApplicationController
+  include RedirectToIdpQuestion
+
   def index
     @idp = decorated_idp
     @form = InterstitialQuestionForm.new({})
     render 'redirect_to_idp_question/redirect_to_idp_question_LOA2'
   end
 
-  def continue
-    @form = InterstitialQuestionForm.new(params['interstitial_question_form'] || {})
-    if @form.valid?
-      if @form.is_yes_selected?
-        selected_answer_store.store_selected_answers('interstitial', @form.selected_answers)
-        redirect_to redirect_to_idp_warning_path
-      else
-        redirect_to idp_wont_work_for_you_one_doc_path
-      end
-    else
-      @idp = decorated_idp
-      flash.now[:errors] = @form.errors.full_messages.join(', ')
-      render 'redirect_to_idp_question/redirect_to_idp_question_LOA2'
-    end
+  def interstitial_selected_path
+    redirect_to_idp_warning_path
   end
 
-  def idp_wont_work_for_you
-    @idp = decorated_idp
-    render 'redirect_to_idp_question/idp_wont_work_for_you'
+  def interstitial_not_selected_path
+    idp_wont_work_for_you_one_doc_path
   end
 
-private
-
-  def decorated_idp
-    @decorated_idp ||= IDENTITY_PROVIDER_DISPLAY_DECORATOR.decorate(selected_identity_provider)
+  def invalid_interstial_path
+    'redirect_to_idp_question/redirect_to_idp_question_LOA2'
   end
 end
